refactor out buildChart functions
